% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accurate_fetcher.r
\name{get.accurate.omdb.info}
\alias{get.accurate.omdb.info}
\title{Fetch movie information from OMDB API by ID or title}
\usage{
get.accurate.omdb.info(
  id = NULL,
  title = NULL,
  type = NULL,
  year = NULL,
  plot = "short"
)
}
\arguments{
\item{id}{Optional IMDb ID (e.g. "tt1285016")}

\item{title}{Optional movie title to search for}

\item{type}{Optional type of result ("movie", "series", or "episode")}

\item{year}{Optional year of release}

\item{plot}{Optional plot length ("short" or "full")}
}
\value{
Movie data from OMDB API
}
\description{
Fetch movie information from OMDB API by ID or title
}
