% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accurate_fetcher.r
\name{get.batch.accurate.omdb.info}
\alias{get.batch.accurate.omdb.info}
\title{Fetch movie information from OMDB API}
\usage{
get.batch.accurate.omdb.info(ids)
}
\arguments{
\item{ids}{IMDb IDs of the movies (e.g. c("tt1285016", "tt1285017"))}
}
\value{
List containing movie information from OMDB API
}
\description{
Fetch movie information from OMDB API
}
