% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rough_fetcher.r
\name{get.rough.omdb.info}
\alias{get.rough.omdb.info}
\title{Fetch movie information from OMDB API by partial title and additional filters}
\usage{
get.rough.omdb.info(
  title,
  type = NULL,
  year = NULL,
  min_rating = NULL,
  genre = NULL
)
}
\arguments{
\item{title}{Partial movie title to search for}

\item{type}{Optional type of result ("movie", "series", or "episode")}

\item{year}{Optional year of release}

\item{min_rating}{Optimal minimum rating for the movie}

\item{genre}{Optional genre filter}
}
\value{
A data frame with movies matching the search criteria
}
\description{
Fetch movie information from OMDB API by partial title and additional filters
}
