% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_api_key.r
\name{set.omdb.api.key}
\alias{set.omdb.api.key}
\title{Set OMDB API key for making requests}
\usage{
set.omdb.api.key(api_key)
}
\arguments{
\item{api_key}{Your OMDB API key string}
}
\description{
Set OMDB API key for making requests
}
